move span validation to its own module
